scaffold whoami command
