+++ ol/cli/Cargo.toml
@@ -65,6 +65,7 @@ bcs = "0.1.2"
 move-vm-types = { path = "../../language/move-vm/types" } # 0L todo: which move-vm/ path?
 move-binary-format = { path = "../../language/move-binary-format" }
 dialoguer = "0.8.0"
+diem-crypto = { path = "../../crypto/crypto" }
 
 
 # Parity with Storage create
